Join to JOIN (#25)
For the upcoming Brigade Congress, I received a note to review presentation slides and noticed a few details on the project description:

JOIN uses all-caps for their name: https://joinpdx.org/

And rather than repeating the word current, I did a line edit so that current is used once. Then I simplified the action verb in the first line so it's more concise.